Rename action build-scan-publish-maven to maven-publish-build-scan
