+++ package.json
@@ -25,7 +25,7 @@
     "package-build-scan-maven-capture-extension": "mvn clean package -f maven-build-scan-capture-extension/pom.xml",
     "package-build-scan-maven-setup-main": "ncc build build-scan-setup-maven/src/main.ts --out dist/build-scan-setup-maven/main",
     "package-build-scan-maven-setup-post": "ncc build build-scan-setup-maven/src/post.ts --out dist/build-scan-setup-maven/post",
-    "package-build-scan-maven-publish": "ncc build build-scan-publish-maven/src/main.ts --out dist/build-scan-publish-maven",
+    "package-build-scan-maven-publish": "ncc build maven-publish-build-scan/src/main.ts --out dist/maven-publish-build-scan",
     "check": "npm-run-all --parallel format:write format:test lint",
     "package": "npm-run-all --parallel package-build-scan-*",
     "build": "npm run check && npm run package",
